add IncomeStatSuite components
