+++ src/components/IncomeStatsSuite/index.tsx
@@ -1,0 +1,5 @@
+export { default as IncomeStat } from './IncomeStat';
+export { default as IncomeStatsCard } from './IncomeStatsCard';
+export { default as IncomeStatsCardHeading } from './IncomeStatsCardHeading';
+export { default as IncomeStatsCardMini } from './IncomeStatsCardMini';
+export { default as IncomeStatsCardPeriodical } from './IncomeStatsCardPeriodical';
